requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask_cors
dotenv
PyPDF2
langchain
langchain_google_genai
langchain_community
google.generativeai
faiss-cpu
